+++ CSM DevKit/CSM DevKit/CHANGELOG.md
@@ -1,5 +1,20 @@
 ﻿# CSM DEV KIT
 
+## 0.0.3 - 29.07-2026
+
+### Added
+
+- Support for [db] commands use prefered database with option [--database] or [-d] indicating the database package name to use.
+
+### Dependencies
+
+| Package                                 | Previous Version | New Version     |
+|:----------------------------------------|:----------------:|:---------------:|
+| CSM.Database.Core                       | -.-.-            | 7.0.0           |
+| CSM.Foundation.Core                     | -.-,-            | 4.0.0           |
+| Spectre.Console						  | -.-.-            | 0.57.2          |
+| System.CommandLine					  | -.-.-			 | 2.0.10		   |
+
 ## 0.0.1 - 28.07-2026
 
 ### Added